# src/main/resources/openapi/api.yaml
openapi: 3.0.3
info:
  title: SpringTest API
  version: 1.0.0
paths:
  /users/{id}:
    get:
      summary: Get user by id
      operationId: getUserById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      summary: Update user
      operationId: updateUserById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete user
      operationId: deleteUserById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Deleted
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users:
    get:
      summary: Get users list
      operationId: getUsers
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserResponse'
                  page:
                    type: integer
                  size:
                    type: integer
                  totalElements:
                    type: integer
                    format: int64
    post:
      summary: Create user
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /courses/{id}:
    get:
      summary: Get course by id
      operationId: getCourseById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Course found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseResponse'
        '404':
          description: Course not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      summary: Update course
      operationId: updateCourseById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseUpdateRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Course not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete course
      operationId: deleteCourseById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Deleted
        '404':
          description: Course not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /courses:
    get:
      summary: Get courses list
      operationId: getCourses
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        '200':
          description: List of courses
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/CourseResponse'
                  page:
                    type: integer
                  size:
                    type: integer
                  totalElements:
                    type: integer
                    format: int64
    post:
      summary: Create course
      operationId: createCourse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseCreateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        profile:
          $ref: '#/components/schemas/ProfileResponse'
    UserCreateRequest:
      type: object
      required:
        - username
      properties:
        username:
          type: string
          minLength: 2
          maxLength: 50
          pattern: '^[a-zA-Z0-9_]+$'
        profile:
          $ref: '#/components/schemas/ProfileCreateRequest'
    UserUpdateRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 2
          maxLength: 50
          pattern: '^[a-zA-Z0-9_]+$'
        profile:
          $ref: '#/components/schemas/ProfileUpdateRequest'
    ProfileResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        bio:
          type: string
    ProfileCreateRequest:
      type: object
      required:
        - bio
      properties:
        bio:
          type: string
          minLength: 1
          maxLength: 500
    ProfileUpdateRequest:
      type: object
      properties:
        bio:
          type: string
          minLength: 1
          maxLength: 500
...
    CourseResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        students:
          type: array
          items:
            $ref: '#/components/schemas/StudentResponse'
    CourseCreateRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 100
        students:
          type: array
          items:
            $ref: '#/components/schemas/StudentCreateRequest'
    CourseUpdateRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 100
        students:
          type: array
          maxItems: 200
          items:
            $ref: '#/components/schemas/StudentUpdateRequest'
    StudentResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
    StudentCreateRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 50
    StudentUpdateRequest:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          minLength: 2
          maxLength: 50
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        code:
          type: string
        timestamp:
          type: string
          format: date-time
        details:
          type: array
          items:
            type: string